# perf: use slices.sort